+++ src/wagtail_jev/urls.py
@@ -6,4 +6,5 @@ app_name = "wagtail_jev"
 
 urlpatterns = [
     path("suggest/", views.suggest, name="suggest"),
+    path("rate/", views.rate, name="rate"),
 ]